requirements.txt
pandas==1.4.4
pymongo==4.2.0
fastapi==0.83.0
python-jose==3.3.0
uvicorn==0.18.3
python-multipart==0.0.5
Jinja2==3.1.2
bcrypt==4.0.0
starlette==0.19.1
PyYAML==6.0
python-dotenv==0.21.0
pymongo[srv]
passlib==1.7.4
notebook
Pillow==9.2.0
deepface==0.0.75
azure-storage-blob==12.13.1
dill==0.3.5.1